feat(account-code): enhance AccountCodeCombobox with add account functionality and update JournalEntryPage to use it
